requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
firebase-admin==6.6.0
Flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.5.0
gcloud==0.18.3
google-api-core==2.24.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jws==0.1.3
lxml==5.3.0
MarkupSafe==3.0.2
MouseInfo==0.1.3
msgpack==1.1.0
multidict==6.1.0
oauth2client==4.1.3
oauthlib==3.2.2
pillow==11.0.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodome==3.21.0
PyGetWindow==0.0.9
PyJWT==2.10.1
PyMsgBox==1.0.9
pyparsing==3.2.0
pyperclip==1.9.0
Pyrebase4==4.8.0
PyRect==0.2.0
PyScreeze==1.0.1
python-docx==1.1.2
python-jwt==2.0.1
pytweening==1.2.0
pywhatkit==5.4
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rsa==4.9
setuptools==75.6.0
six==1.17.0
soupsieve==2.6
twilio==9.4.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
Werkzeug==3.1.3
wikipedia==1.4.0
yarl==1.18.3